+++ tests/runner/testresource_test.go
@@ -1,4 +1,9 @@
-package utils
+// ------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+// ------------------------------------------------------------
+
+package runner
 
 import (
 	"fmt"
